[FEAT] Separate the module for the initial situation where the nonce is 0

+++ srcs/requirements/django/result/deploy.py
@@ -10,14 +10,25 @@ class TournamentResultManager:
         self.solc_version = "0.6.0"
         self.chain_address = os.getenv("CHAIN_ADDRESS")
         self.private_key = os.getenv("PRIVATE_KEY")
+        if self.chain_address == None or self.private_key == None:
+            print("Please set the CHAIN_ADDRESS and PRIVATE_KEY in the .env file.")
+            return
         self.w3 = Web3(Web3.HTTPProvider(provider))
-        print(self.w3.eth.get_transaction_count(self.chain_address))
         self.chain_id = 1337
-        receipt = self.__deploy_contract(sol_path)
-        if receipt == None:
-            self.contract_address = os.getenv("CONTRACT_ADDRESS")
+        
+        if self.w3.eth.get_transaction_count(self.chain_address) == 0 or os.getenv("CONTRACT_ADDRESS") == "":
+            self.__set_initial_settings(sol_path)
         else:
-            self.contract_address = str(receipt.contractAddress)
+            with open("./abi.json", 'r', encoding='utf-8') as file:
+                self.abi = file.read()
+            self.contract_address = os.getenv("CONTRACT_ADDRESS")
+
+    def __set_initial_settings(self, sol_path):
+        bytecode = self.__compile_sol(sol_path)
+        contract_address = self.__deploy_contract(bytecode)
+        self.contract_address = str(contract_address)
+        print("The contract address is as follows: " + self.contract_address)
+        print("Set the above contract address as the CONTRACT_ADDRESS in the .env file.")
 
     def __compile_sol(self, sol_path):
         install_solc(self.solc_version)
@@ -37,33 +48,31 @@ class TournamentResultManager:
             },
             solc_version=self.solc_version,
         )
-        bytecode = compiled_sol["contracts"]["TournamentContract.sol"]["TournamentContract"]["evm"]["bytecode"]["object"]
         self.abi = compiled_sol["contracts"]["TournamentContract.sol"]["TournamentContract"]["abi"]
+        bytecode = compiled_sol["contracts"]["TournamentContract.sol"]["TournamentContract"]["evm"]["bytecode"]["object"]
+
         # for get abi
-        with open("abi.json", "w") as file:
-            json.dump(self.abi, file)
+        if not os.path.exists("abi.json"):
+            with open("abi.json", "w") as file:
+                json.dump(self.abi, file)
         return bytecode
 
-    def __deploy_contract(self, sol_path):
+    def __deploy_contract(self, bytecode):
         nonce = self.w3.eth.get_transaction_count(self.chain_address)
-        if nonce != 0:
-            self.abi = os.getenv("ABI")
-            return None
-        bytecode = self.__compile_sol(sol_path)
+
         Tournament = self.w3.eth.contract(abi=self.abi, bytecode=bytecode)
         transaction = Tournament.constructor().build_transaction(
             {
                 "chainId": self.chain_id,
                 "gasPrice": self.w3.eth.gas_price,
                 "from": self.chain_address,
-                "nonce": nonce,
+                "nonce": nonce
             }
         )
         signed_txn = self.w3.eth.account.sign_transaction(
             transaction, private_key=self.private_key)
         tx_hash = self.w3.eth.send_raw_transaction(signed_txn.rawTransaction)
         tx_receipt = self.w3.eth.wait_for_transaction_receipt(tx_hash)
-        print(tx_receipt.contractAddress)
         return tx_receipt.contractAddress
 
     def start_game(self, id, timestamp, players):
@@ -75,7 +84,7 @@ class TournamentResultManager:
                 "chainId": self.chain_id,
                 "gasPrice": self.w3.eth.gas_price,
                 "from": self.chain_address,
-                "nonce": nonce,
+                "nonce": nonce
             }
         )
         signed_txn = self.w3.eth.account.sign_transaction(
@@ -121,10 +130,10 @@ load_dotenv()
 tournament_contract = TournamentResultManager(
     "../../blockchain/TournamentContract.sol", os.getenv("GANACHE_URL"))
 
-tournament_contract.start_game(20, 1625940800, [263, 456, 989, 1011])
-tournament_contract.save_sub_game(20, [2, 10, 1])
-tournament_contract.save_sub_game(20, [3, 2, 10])
-tournament_contract.save_sub_game(20, [1, 4, 10])
+tournament_contract.start_game(4, 1625940800, [263, 456, 989, 1011])
+tournament_contract.save_sub_game(4, [2, 10, 1])
+tournament_contract.save_sub_game(4, [3, 2, 10])
+tournament_contract.save_sub_game(4, [1, 4, 10])
 
 a = tournament_contract.get_all_tournaments()
 print(a)
